+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 2.x.x - 2021-05-27
+
+### Added 
+- Width prop in QasTooltip
+
+### Fixed
+- Fix import filterObject in helpers
 ## 2.2.0 - 2021-05-26
 
 ### Added